requirements.txt
Flask
gunicorn
tensorflow
numpy
python-telegram-bot
requests